[+] Mostly implement instance data uploading using SSBOs. Correctly wait for in-flight data to be freed using the swap-chain.
[*]
Introduced a horrible bug where only one frame has the correct SSBO data.

+++ VulkanRenderingTest/Resources.h
@@ -99,7 +99,7 @@ struct DrawCall
 	/*
 	 * The amount of mesh instances in the m_pMeshInstances array.
 	 */
-	size_t m_NumInstances = 0;
+	uint32_t m_NumInstances = 0;
 
 	/*
 	 * If true, this geometry is drawn in a separate forward pass after the deferred stage ends.
